add query_raw_txt for transaction
